BREAKING CHANGE: Remove shaded dependencies and use HttpClient interface (#219)
* BREAKING CHANGE: Remove shaded dependencies

* feat: Use HttpClient interface

* test: Fix tests

* fix: Fix compile errors

* fix: Fix all tests

* fix: Fix NPE

* test: Add more tests

* test: Add more tests and doc comments

* test: More tests and minor refactoring

+++ src/main/java/com/spotify/github/v3/issues/Issue.java
@@ -67,7 +67,7 @@ public interface Issue extends CloseTracking {
 
   /** Number. */
   @Nullable
-  Integer number();
+  Long number();
 
   /** Indicates the state of the issues to return. Can be either open, closed, or all. */
   @Nullable
